chore(core): extract & remove some legacy components [#67]
